Heap max min

+++ general/types.h
@@ -4,7 +4,8 @@
 using T1 = int;
 using T  = unsigned int;
 using TT = unsigned long long;
-using Size = int;
+using Size = size_t;
+//using Size = int;
 
 using ContainerElemType = int;
 using ContainerRange = int;
